requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0  # For headless browser automation
selenium>=4.15.0  # Optional: alternative to Playwright
rich>=13.0.0  # For beautiful terminal output and progress bars
psycopg2-binary>=2.9.0  # PostgreSQL adapter for Python
redis>=5.0.0  # Redis client for caching

# Server dependencies (for downloader_server.py)
fastapi>=0.104.0  # Web framework for API server
uvicorn[standard]>=0.24.0  # ASGI server
pydantic>=2.0.0  # Data validation
pydantic-settings>=2.0.0  # Settings management
python-multipart>=0.0.6  # For form data parsing
python-dateutil>=2.8.0  # For date manipulation
requests>=2.31.0  # For Telegram notifications